+++ node/resolvers/search/facets.ts
@@ -1,8 +1,30 @@
+import {
+  formatTranslatableStringV2
+} from '@vtex/api'
+
 export const resolvers = {
   Facet: {
-    values: (facet: Facet, args: FacetValuesArgs) => {
+    name: (facet: Facet, _: FacetValuesArgs, ctx: Context) => {
+      const { locale, tenant } = ctx.vtex
+      const from = locale ?? tenant?.locale
+      return formatTranslatableStringV2({
+        content: facet.name,
+        from,
+      })
+    },
+    values: (facet: Facet, args: FacetValuesArgs, ctx: Context) => {
       const { values } = facet
-      return values.slice(args.from ?? 0, args.to)
+      const { locale, tenant } = ctx.vtex
+      const from = locale ?? tenant?.locale
+      return values.slice(args.from ?? 0, args.to).map(value => {
+        return {
+          ...value,
+          name: formatTranslatableStringV2({
+            content: value.name,
+            from,
+          })
+        }
+      })
     },
     quantity: (facet: Facet) => {
       return facet.values.length ?? 0