+++ node/resolvers/search/index.ts
@@ -251,6 +251,7 @@ export const fieldResolvers = {
   ...productPriceRangeResolvers,
 }
 
+// TODO: We probably don't need this anymore. I think it is done on the rewriter.
 export const getCompatibilityArgs = async <T extends QueryArgs>(
   ctx: Context,
   args: T
@@ -407,11 +408,7 @@ export const queries = {
       args
     )) as FacetsInput
 
-    let { fullText, searchState } = args
-
-    if (fullText) {
-      fullText = await translateToStoreDefaultLanguage(ctx, args.fullText!)
-    }
+    const { fullText, searchState } = args
 
     const {
       clients: { biggySearch, search, checkout, vbase },